MultiSelectionUpdate - extracted common models between extension and UI
