+++ src-common/models/package-source.ts
@@ -1,0 +1,5 @@
+export interface PackageSource {
+  name: string;
+  url: string;
+  authorizationHeader: string | undefined;
+}
