--- src-ext/models/project.ts
@@ -1,10 +1,0 @@
-export interface Project {
-  name: string;
-  fsPath: string;
-  packages: Package[];
-}
-
-export interface Package {
-  id: string;
-  version: string;
-}
