+++ src/app/services/nuget-api.service.ts
@@ -6,7 +6,7 @@ import { coerce, prerelease, rcompare } from 'semver';
 import { ApiIndexResponse, ApiResource } from '../models/api-index-response';
 import { PackageRowModel } from '../models/package-row-model';
 import { CatalogEntry, PackageMetaResponse, RegistrationLeaf, RegistrationPage } from '../models/package-meta';
-import { PackageSource } from '../models/package-source';
+import { PackageSource } from '../../../src-common/models/package-source';
 import { PackageSearchResult, SearchResults } from '../models/search-results';
 
 @Injectable({
