+++ src/app/services/package-manager.service.spec.ts
@@ -2,8 +2,8 @@ import { TestBed } from '@angular/core/testing';
 import { of, skip, take, zip } from 'rxjs';
 import { Category } from '../models/category';
 import { PackageRowModel } from '../models/package-row-model';
-import { PackageSource } from '../models/package-source';
-import { Project } from '../models/project';
+import { PackageSource } from '../../../src-common/models/package-source';
+import { Project } from '../../../src-common/models/project';
 import { NuGetApiService } from './nuget-api.service';
 import { PackageManagerService } from './package-manager.service';
 import { VscodeService } from './vscode.service';
