+++ src/app/services/vscode.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { UIMessage } from '../models/ui-message';
+import { UIMessage } from '../../../src-common/models/ui-message';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare function acquireVsCodeApi(): any;
